requirements.txt
psycopg2-binary
Scrapy
scrapy-playwright
asyncio
Twisted
boto3
prefect
prefect-shell
sqlalchemy-bigquery
google-cloud-bigquery
plotly
docstr-coverage
pytest